+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/alarm/AlarmRecord.java
@@ -82,6 +82,7 @@ public class AlarmRecord extends Record {
     private long startTime;
     @Column(name = ALARM_MESSAGE, length = 512)
     @ElasticSearch.MatchQuery
+    @BanyanDB.MatchQuery
     private String alarmMessage;
     @Column(name = RULE_NAME)
     private String ruleName;